pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iocontrol"
version = "0.1.0"
description = "High-performance asynchronous I/O control for embedded systems"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Landon Bell", email = "landon.bell@valorence.com" }]
dependencies = ["asyncio", "typing-extensions>=4.0.0"]

[project.optional-dependencies]
hardware = ["smbus2"]

dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
simulation = [
    "pyyaml>=6.0", # For config files
]

[project.urls]
Homepage = "https://github.com/ValorenceCLE/iocontrol"

[tool.hatch.build.targets.wheel]
packages = ["src/iocontrol"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
